move commands to their own file, split the commands into their own methods
